Add files via upload
The abstract of "Great Food Restaurant" project is developed to reduce the manual work load of the workers by providing automated billing operation.

Thus it greatly reduces the time and improves the overall efficiency of the restaurant.

The project is also designed to view the total number of sales encountered in a day. 

Furthermore, it allows the user to aspect the quantity of food items sold in a day with the help of a pictorial representation by means of bar graph.